requirements.txt
appdirs==1.4.4
asgiref==3.3.1    
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.4
djangorestframework==3.12.2
filelock==3.0.12
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.2
PyMySQL==0.10.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1